corrected migration, rename to finish

+++ database/migrations/2019_02_16_202450_rename_user_recipes_likes_to_likes.php
@@ -14,27 +14,34 @@ class RenameUserRecipesLikesToLikes extends Migration
     public function up()
     {
         if (Schema::hasTable('user_recipe_likes')) {
-            Schema::rename('user_recipe_likes', 'likes');
-            Schema::table('likes', function (Blueprint $table) {
-                $table->dropColumn('user_id');
-                $table->dropColumn('recipe_id');
+            if (!Schema::hasTable('likes')) {
+                Schema::rename('user_recipe_likes', 'likes');
 
-                $table->morphs('likeable');
-                $table->softDeletes();
-                $table->timestamps();
-                $table->foreign('user_id')->references('id')->on('users');
-            });
+                Schema::table('likes', function (Blueprint $table) {
+                    $table->dropColumn('user_id');
+                    $table->dropColumn('recipe_id');
+
+                    $table->unsignedInteger('user_id')->nullable();
+                    $table->morphs('likeable');
+                    $table->softDeletes();
+                    $table->timestamps();
+                    $table->foreign('user_id')->references('id')->on('users');
+                });
+            }
         } else {
-            Schema::create('likes', function (Blueprint $table) {
-                $table->increments('id');
-                $table->unsignedInteger('user_id')->nullable();
+            if (!Schema::hasTable('likes')) {
+                Schema::create('likes', function (Blueprint $table) {
+                    $table->increments('id');
+                    $table->unsignedInteger('user_id')->nullable();
 
-                $table->morphs('likeable');
-                $table->softDeletes();
-                $table->timestamps();
+                    $table->morphs('likeable');
+                    $table->softDeletes();
+                    $table->timestamps();
 
-                $table->foreign('user_id')->references('id')->on('users');
-            });
+                    $table->foreign('user_id')->references('id')->on('users');
+                });
+            }
+
         }
     }
 
@@ -45,8 +52,7 @@ class RenameUserRecipesLikesToLikes extends Migration
      */
     public function down()
     {
-        if (Schema::hasTable('user_recipe_likes')) {
-            Schema::rename('user_recipe_likes', 'likes');
-        }
+        Schema::dropIfExists('user_recipe_likes');
+        Schema::dropIfExists('likes');
     }
 }
